+++ Adjust/Source/Adjust/Public/AdjustDelegates.h
@@ -16,6 +16,7 @@
 #include "AdjustEventSuccess.h"
 #include "AdjustEventFailure.h"
 #include "AdjustPurchaseVerificationResult.h"
+#include "AdjustSkanConversionDataMap.h"
 #include "AdjustDelegates.generated.h"
 
 // subscription multiplatform delegates
@@ -26,6 +27,7 @@ DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnEventSuccessDelegate, const FAdju
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnEventFailureDelegate, const FAdjustEventFailure&, EventFailure);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnDeferredDeeplinkDelegate, const FString&, Deeplink);
 // subscription ios delegates
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnSkanConversionValueUpdatedDelegate, const FAdjustSkanConversionDataMap&, ConversionData);
 
 // one-time multiplatform delegates
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnIsEnabledDelegate, const bool, IsEnabled);
@@ -35,6 +37,7 @@ DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnLastDeeplinkGetterDelegate, const
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnDeeplinkResolutionDelegate, const FString&, ResolvedLink);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnLinkResolutionDelegate, const FString&, ResolvedLink);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnSdkVersionGetterDelegate, const FString&, SdkVersion);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnPurchaseVerificationFinishedDelegate, const FAdjustPurchaseVerificationResult&, VerificationResult);
 // one-time ios delegates
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnIdfaGetterDelegate, const FString&, Idfa);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnIdfvGetterDelegate, const FString&, Idfv);
@@ -44,8 +47,6 @@ DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnUpdateSkanConversionValueDelegate
 // one-time android delegates
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnGoogleAdIdGetterDelegate, const FString&, GoogleAdId);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnAmazonAdIdGetterDelegate, const FString&, AmazonAdId);
-// purchase verification delegates
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnPurchaseVerificationFinishedDelegate, const FAdjustPurchaseVerificationResult&, VerificationResult);
 
 
 UCLASS(Blueprintable)
@@ -76,6 +77,11 @@ public:
     UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
     FOnDeferredDeeplinkDelegate OnDeferredDeeplinkDelegate;
 
+    // subscription ios delegates
+
+    UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
+    FOnSkanConversionValueUpdatedDelegate OnSkanConversionValueUpdatedDelegate;
+
     // one-time multiplatform delegates
 
     UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
@@ -98,6 +104,9 @@ public:
 
     UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
     FOnSdkVersionGetterDelegate OnSdkVersionGetterDelegate;
+
+    UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
+    FOnPurchaseVerificationFinishedDelegate OnPurchaseVerificationFinishedDelegate;
 
     // one-time ios delegates
 
@@ -123,9 +132,4 @@ public:
 
     UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
     FOnAmazonAdIdGetterDelegate OnAmazonAdIdGetterDelegate;
-
-    // purchase verification delegates
-
-    UPROPERTY(BlueprintAssignable, BlueprintCallable, Category = Adjust)
-    FOnPurchaseVerificationFinishedDelegate OnPurchaseVerificationFinishedDelegate;
 };