+++ Adjust/Source/Adjust/Public/AdjustPlayStorePurchase.h
@@ -2,8 +2,8 @@
 //  AdjustPlayStorePurchase.h
 //  Adjust SDK
 //
-//  Created by Uglješa Erceg (@uerceg) on 27th September 2018.
-//  Copyright © 2018-Present Adjust GmbH. All rights reserved.
+//  Created by Uglješa Erceg (@uerceg) on 6th November 2025.
+//  Copyright © 2025-Present Adjust GmbH. All rights reserved.
 //
 
 #pragma once